requirements.txt
streamlit
pandas
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
gspread
google-cloud-aiplatform
vertexai